Fixed collision to be more precise
Before there was an issue that when you were up against a wall you couldn't move the final pixel to touch the ground. This was likely caused by discrepancies in place_meeting and collision_point where once it realized you would hit the ground it would see you touching the wall as an excuse to not move down any more.